FIX: Narrow precision in the timer class resulted in deltatime = 0.

+++ Engine/include/core/engine.h
@@ -20,7 +20,7 @@ namespace Core
 	{
 	public:
 		void Initialize();
-		void Update(float);
+		void Update(double);
 		void Render();
 		void ShutDown();
