+++ Engine/include/core/timer.h
@@ -13,11 +13,11 @@ enum Unit
 class Timer
 {
 public:
-	float GetDeltaTime(Unit unit);
+	long long GetDeltaTime(Unit unit);
 
 private:
 	template <typename T>
-	float ReturnType(std::chrono::nanoseconds deltaTime);
+	long long ReturnType(std::chrono::nanoseconds deltaTime);
 
 	std::chrono::high_resolution_clock m_timer;
 	std::chrono::nanoseconds m_lastTimePoint = std::chrono::duration<long long, std::nano>(0);
@@ -25,36 +25,36 @@ private:
 
 // Base
 template<typename T>
-float Timer::ReturnType(std::chrono::nanoseconds deltaTime)
+long long Timer::ReturnType(std::chrono::nanoseconds deltaTime)
 {
 	throw std::runtime_error("Type is not part of std::chrono");
 }
 
 // Specializations
 template<>
-inline float Timer::ReturnType<std::chrono::nanoseconds>(std::chrono::nanoseconds deltaTime)
+inline long long Timer::ReturnType<std::chrono::nanoseconds>(std::chrono::nanoseconds deltaTime)
 {
-	return static_cast<float>(deltaTime.count());
+	return static_cast<long long>(deltaTime.count());
 }
 
 template<>
-inline float Timer::ReturnType<std::chrono::microseconds>(std::chrono::nanoseconds deltaTime)
+inline long long Timer::ReturnType<std::chrono::microseconds>(std::chrono::nanoseconds deltaTime)
 {
 	std::chrono::microseconds value = std::chrono::duration_cast<std::chrono::microseconds>(deltaTime);
-	return static_cast<float>(value.count());
+	return static_cast<long long>(value.count());
 }
 
 template<>
-inline float Timer::ReturnType<std::chrono::milliseconds>(std::chrono::nanoseconds deltaTime)
+inline long long Timer::ReturnType<std::chrono::milliseconds>(std::chrono::nanoseconds deltaTime)
 {
 	std::chrono::milliseconds value = std::chrono::duration_cast<std::chrono::milliseconds>(deltaTime);
-	return static_cast<float>(value.count());
+	return static_cast<long long>(value.count());
 }
 
 template<>
-inline float Timer::ReturnType<std::chrono::seconds>(std::chrono::nanoseconds deltaTime)
+inline long long Timer::ReturnType<std::chrono::seconds>(std::chrono::nanoseconds deltaTime)
 {
 	std::chrono::seconds value = std::chrono::duration_cast<std::chrono::seconds>(deltaTime);
-	return static_cast<float>(value.count());
+	return static_cast<long long>(value.count());
 }
 //
